created dictionary file called industry_dict as well as prospicio
